Implement adding of 2 roman numerals
Note: In the conversion tests we don't need to (and don't want to)
test all possible combinations of inputs and outputs. Each test
asserts a certain functionality, and documents it. By adding the
ability to add 2 roman numerals together we need to increase the
range of supported inputs and outputs, and thus test the conversion
implicitly, but the existing conversion tests stay valid. Adding
new tests often requires existing, already tested functionality to
be changed or extended, but rarely should it be required to change
existing tests to add new tests.