Refactor roman numeral conversion test for simplicity
Note:
Unit tests can be safely refactored by adding redundancy, thus always
keeping the functionality tested during the refactoring of the test.
In the next commit, the old version of the test will be removed and
only the simpler, refactored version will be kept. This does not
necessarily have to be split into 2 commits, but we do it in this
case to demonstrate what the steps involved in refactoring of unit
tests may look like.