Simplify conversion to roman strings in Java example solution
This is now more similar to the Kotlin example solution, and overall
shorter than before. After the opposite conversion (i.e. parsing
from roman numeral string) is is also simplified accordingly, the
code will also in general be a bit more normalized (i.e. it will
contain less duplicated information, which is not to be confused
with "less duplicated code").

This refactoring is safe due to the unit test coverage, but would be
quite risky and error prone without unit tests.